+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 9.0
 - CSS add length to previous meals select in DailyMenus.php
 - CSS FlatSIS fix calendar menu text wrapping in stylesheet.css
 - Add Export list button in TransactionsReport.php
+- Add Food Service icon to list in ServeMenus.php
 
 Changes in 8.9
 --------------